added sort post by topic and corrected create post
