+++ apps/api/src/controllers/post.ts
@@ -18,6 +18,33 @@ class postController {
       next(createHttpError(error.status, error.message));
     }
   }
+  static async createPost(req: Request, res: Response, next: NextFunction) {
+    try {
+      const result = await postService.createPost(req.body);
+      res.json({
+        status: 200,
+        message: "Post created successfully",
+        data: result,
+      });
+    } catch (e) {
+      const error = e as ErrorWithStatus;
+      next(createHttpError(error.status, error.message));
+    }
+  }
+  static async postsByTopicId(req: Request, res: Response, next: NextFunction) {
+    console.log(req.params["topicId"])
+    try {
+      const result = await postService.postsByTopicId({topicId: req.params["topicId"]});
+      res.json({
+        status: 200,
+        message: "Posts by ID fetched successfully",
+        data: result,
+      });
+    } catch (e) {
+      const error = e as ErrorWithStatus;
+      next(createHttpError(error.status, error.message));
+    }
+  }
 
   static async postById(req: Request, res: Response, next: NextFunction) {
     try {
