+++ apps/api/src/controllers/topic.ts
@@ -33,35 +33,6 @@ class topicController {
       next(createHttpError(error.status, error.message));
     }
   }
-  static async createPost(req: Request, res: Response, next: NextFunction) {
-    try {
-      const result = await topicService.createPost(req.body);
-      res.json({
-        status: 200,
-        message: "Post created successfully",
-        data: result,
-      });
-    } catch (e) {
-      const error = e as ErrorWithStatus;
-      next(createHttpError(error.status, error.message));
-    }
-  }
-
-  static async postsByTopicId(req: Request, res: Response, next: NextFunction) {
-    try {
-      const result = await topicService.postsBytopicId({
-        id: req.params["id"],
-      });
-      res.json({
-        status: 200,
-        message: "Posts by ID fetched successfully",
-        data: result,
-      });
-    } catch (e) {
-      const error = e as ErrorWithStatus;
-      next(createHttpError(error.status, error.message));
-    }
-  }
 }
 
 export { topicController };
