+++ apps/api/src/routes/post.ts
@@ -7,5 +7,7 @@ const router = Router();
 
 router.get("/", authMiddleware, postController.allPosts);
 router.get("/:id", postController.postById);
+router.get("/:topicId/posts", authMiddleware, postController.postsByTopicId);
+router.post("/:topicId/posts", authMiddleware, postController.createPost);
 
 export default router;
