+++ apps/api/src/routes/topic.ts
@@ -6,8 +6,6 @@ import { authMiddleware } from "../middleware/auth";
 const router = Router();
 
 router.get("/", authMiddleware, topicController.allTopics);
-router.post("/:id/posts", authMiddleware, topicController.createPost);
-router.get("/:id", authMiddleware, topicController.topicById);
-router.get("/:id/posts", authMiddleware, topicController.postsByTopicId);
+router.get("/:id", authMiddleware, topicController.topicById); 
 
 export default router;
